add: message validation mechanism
