+++ services/pong/app/src/handlers/pong.ts
@@ -1,25 +1,14 @@
 import type { Game } from '../classes/game.class.js';
 import { updatePaddlePos } from './paddle.js';
 import type { Player } from '../classes/player.class.js';
-
-export interface keysObj {
-	_w: boolean,
-	_s: boolean,
-	_ArrowUp: boolean,
-	_ArrowDown: boolean,
-	[key: string]: boolean,
-}
-
-export interface messObj {
-	_keys: keysObj,
-	_timeStamp: number,
-}
+import { validMess, type messObj, type keysObj } from './mess.validation.js';
 
 export function setUpGame(game: Game) {
 	if (!game.players[0] || !game.players[1])
 		return; //TODO: deal with that
 	const player1: Player = game.players[0];
 	const player2: Player = game.players[1];
+	
 	if (game.local)
 		setMessEvent(player1, player2, local);
 	else {
@@ -30,8 +19,12 @@ export function setUpGame(game: Game) {
 
 export function setMessEvent(player: Player, opponent: Player, sendReply: Function) {
 	let lastFrameTime: number = 0;
-	player.socket.on("message", (payload: any) => {
-		const mess: messObj = JSON.parse(payload);
+	player.socket.on("message", (payload: string) => {
+		let mess: messObj;
+		try { mess = JSON.parse(payload); }
+		catch (err) { return; };
+		if (!validMess(mess))
+			return;
 		const keys: keysObj = mess._keys;
 		const delta: number = mess._timeStamp - lastFrameTime;
 		lastFrameTime = mess._timeStamp;
